+++ aiida/tools/archive/create.py
@@ -36,7 +36,7 @@ from aiida.tools.graph.graph_traversers import get_nodes_export, validate_traver
 from .abstract import ArchiveFormatAbstract, ArchiveWriterAbstract
 from .common import batch_iter, entity_type_to_orm
 from .exceptions import ArchiveExportError, ExportValidationError
-from .implementations.sqlite import ArchiveFormatSqlZip
+from .implementations.sqlite_zip import ArchiveFormatSqlZip
 
 __all__ = ('create_archive', 'EXPORT_LOGGER')
