+++ aiida/tools/archive/implementations/sqlite_zip/__init__.py
