+++ aiida/tools/archive/implementations/sqlite_zip/common.py
