+++ aiida/tools/archive/implementations/sqlite_zip/reader.py
