+++ aiida/tools/archive/implementations/sqlite_zip/writer.py
