+++ tests/tools/archive/migration/test_legacy_funcs.py
@@ -12,8 +12,8 @@
 import pytest
 
 from aiida import get_version
-from aiida.tools.archive.implementations.sqlite.migrations.legacy import LEGACY_MIGRATE_FUNCTIONS
-from aiida.tools.archive.implementations.sqlite.migrations.utils import verify_metadata_version
+from aiida.storage.sqlite_zip.migrations.legacy import LEGACY_MIGRATE_FUNCTIONS
+from aiida.storage.sqlite_zip.migrations.legacy.utils import verify_metadata_version
 from tests.utils.archives import get_archive_file, read_json_files
