+++ tests/tools/archive/migration/test_v04_to_v05.py
@@ -8,7 +8,7 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """Test archive file migration from export version 0.4 to 0.5"""
-from aiida.tools.archive.implementations.sqlite.migrations.legacy import migrate_v4_to_v5
+from aiida.storage.sqlite_zip.migrations.legacy import migrate_v4_to_v5
 
 
 def test_migrate_external(migrate_from_func):
