+++ tests/tools/archive/migration/test_v07_to_v08.py
@@ -8,10 +8,7 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """Test archive file migration from export version 0.7 to 0.8"""
-from aiida.tools.archive.implementations.sqlite.migrations.legacy.v07_to_v08 import (
-    migrate_v7_to_v8,
-    migration_default_link_label,
-)
+from aiida.storage.sqlite_zip.migrations.legacy.v07_to_v08 import migrate_v7_to_v8, migration_default_link_label
 
 
 def test_migrate_external(migrate_from_func):
