+++ tests/tools/archive/migration/test_v08_to_v09.py
@@ -8,10 +8,7 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 """Test archive file migration from export version 0.8 to 0.9"""
-from aiida.tools.archive.implementations.sqlite.migrations.legacy.v08_to_v09 import (
-    migrate_v8_to_v9,
-    migration_dbgroup_type_string,
-)
+from aiida.storage.sqlite_zip.migrations.legacy.v08_to_v09 import migrate_v8_to_v9, migration_dbgroup_type_string
 
 
 def test_migrate_external(migrate_from_func):
